feat(auth-service): Restructured auth service by ddd and hexagonal
