--- auth-service/src/main/java/com/tanveer/authservice/config/CustomAuthenticationEntryPoint.java
@@ -1,28 +1,0 @@
-package com.tanveer.authservice.config;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.tanveer.authservice.dto.ApiResponse;
-import io.jsonwebtoken.io.IOException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.OutputStream;
-
-@Component
-class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-  @Override
-  public void commence(HttpServletRequest request, HttpServletResponse response,
-                       AuthenticationException authException) throws IOException, java.io.IOException {
-    ApiResponse re = new ApiResponse(false,"Authentication require",null);
-    response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-    response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-    OutputStream responseStream = response.getOutputStream();
-    ObjectMapper mapper = new ObjectMapper();
-    mapper.writeValue(responseStream, re);
-    responseStream.flush();
-  }
-
-}
