--- auth-service/src/main/java/com/tanveer/authservice/config/JwtAuthorizationFilter.java
@@ -1,82 +1,0 @@
-package com.tanveer.authservice.config;
-
-import com.tanveer.authservice.repository.TokenRepository;
-import com.tanveer.authservice.service.JwtService;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.lang.NonNull;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthorizationFilter extends OncePerRequestFilter {
-
-  private HandlerExceptionResolver exceptionResolver;
-
-  private final JwtService jwtService;
-
-  private final UserDetailsService userDetailsService;
-
-  private final TokenRepository tokenRepository;
-
-  @Autowired
-  public JwtAuthorizationFilter(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver exceptionResolver,
-                                JwtService jwtService, UserDetailsService userDetailsService, TokenRepository tokenRepository) {
-    this.exceptionResolver = exceptionResolver;
-    this.jwtService = jwtService;
-    this.userDetailsService = userDetailsService;
-    this.tokenRepository = tokenRepository;
-  }
-
-  @Override
-  protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response,
-                                  @NonNull FilterChain filterChain) throws ServletException, IOException {
-
-    final String authHeader = request.getHeader("Authorization");
-    final String jwt;
-    final String userEmail;
-
-    try {
-      if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-        filterChain.doFilter(request, response);
-        return;
-      }
-
-      jwt = authHeader.substring(7);
-      userEmail = jwtService.extractUsername(jwt);
-
-      if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-        UserDetails userDetails = userDetailsService.loadUserByUsername(userEmail);
-        var isTokenValid = tokenRepository.findByToken(jwt)
-          .map(t -> !t.isExpired() && !t.isRevoked())
-          .orElse(false);
-        if (jwtService.isTokenValid(jwt, userDetails) && isTokenValid) {
-          UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
-            userEmail,
-            null,
-            userDetails.getAuthorities()
-          );
-          authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-          SecurityContextHolder.getContext().setAuthentication(authToken);
-        }
-      }
-      filterChain.doFilter(request, response);
-    } catch (Exception ex) {
-      exceptionResolver.resolveException(request, response, null, ex);
-    }
-  }
-}
