--- auth-service/src/main/java/com/tanveer/authservice/domain/TokenType.java
@@ -1,5 +1,0 @@
-package com.tanveer.authservice.domain;
-
-public enum TokenType {
-  BEARER
-}
