+++ auth-service/src/main/java/com/tanveer/authservice/domain/User.java
@@ -1,0 +1,17 @@
+package com.tanveer.authservice.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public final class User {
+    private final String id;
+    private final String email;
+    private final String password;
+    private final String role;
+
+    public static User create(String id, String email, String password, String role) {
+        return new User(id,email,password,role);
+    }
+}
