--- auth-service/src/main/java/com/tanveer/authservice/domain/UserEntity.java
@@ -1,28 +1,0 @@
-package com.tanveer.authservice.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Entity
-@Table(name = "user_entity")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserEntity {
-
-    @Id
-    private String id = UUID.randomUUID().toString();
-
-    @Column(unique = true, nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String role;
-}
