--- auth-service/src/main/java/com/tanveer/authservice/dto/ApiResponse.java
@@ -1,13 +1,0 @@
-package com.tanveer.authservice.dto;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ApiResponse {
-  private boolean success;
-  private String message;
-  private Object data;
-}
