--- auth-service/src/main/java/com/tanveer/authservice/exception/CustomException.java
@@ -1,12 +1,0 @@
-package com.tanveer.authservice.exception;
-
-public class CustomException extends Exception{
-
-  private static final long serialVersionUID = 1L;
-
-  public CustomException(String message) {
-    super(message);
-  }
-
-}
-
