+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/dto/ApiResponse.java
@@ -1,0 +1,13 @@
+package com.tanveer.authservice.infrastructure.dto;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ApiResponse {
+  private boolean success;
+  private String message;
+  private Object data;
+}
