+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/dto/AuthResponse.java
@@ -1,0 +1,20 @@
+package com.tanveer.authservice.infrastructure.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthResponse {
+
+  @JsonProperty("access_token")
+  private String accessToken;
+  @JsonProperty("refresh_token")
+  private String refreshToken;
+}
+
