+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/dto/SignUpRequestDto.java
@@ -1,0 +1,26 @@
+package com.tanveer.authservice.infrastructure.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignUpRequestDto {
+
+    @NotBlank(message = "Email is mandatory")
+    @Pattern(regexp = "^[a-zA-Z0-9_!#$%&amp;'*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$",
+            message = "Email should match the pattern a-z @ a-z .com")
+    private String email;
+
+    @NotBlank(message = "Password is mandatory")
+    private String password;
+
+    @NotBlank(message = "Role is mandatory")
+    private String role;
+}
