+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/exception/CustomException.java
@@ -1,0 +1,12 @@
+package com.tanveer.authservice.infrastructure.exception;
+
+public class CustomException extends Exception{
+
+  private static final long serialVersionUID = 1L;
+
+  public CustomException(String message) {
+    super(message);
+  }
+
+}
+
