+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/TokenEntity.java
@@ -1,0 +1,42 @@
+package com.tanveer.authservice.infrastructure.persistance;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "token")
+public class TokenEntity {
+
+  @Id
+  @GeneratedValue
+  private Integer id;
+
+  @Column(unique = true)
+  public String token;
+
+  @Enumerated(EnumType.STRING)
+  public TokenType tokenType = TokenType.BEARER;
+
+  public boolean revoked;
+
+  public boolean expired;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "user_entity_id")
+  public UserEntity userEntity;
+}
