+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/TokenJpaRepository.java
@@ -1,0 +1,21 @@
+package com.tanveer.authservice.infrastructure.persistance;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface TokenJpaRepository extends JpaRepository<TokenEntity, Integer> {
+
+  @Query(value = """
+      select t from TokenEntity t inner join UserEntity u\s
+      on t.userEntity.email = u.email\s
+      where u.email = :email and (t.expired = false or t.revoked = false)\s
+      """)
+  List<TokenEntity> findAllValidTokenByUser(String email);
+
+  Optional<TokenEntity> findByToken(String tokenEntity);
+}
