+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/TokenRepository.java
@@ -1,0 +1,8 @@
+package com.tanveer.authservice.infrastructure.persistance;
+
+import com.tanveer.authservice.domain.User;
+
+public interface TokenRepository {
+    void revokeAllUserTokens(String email);
+    void saveUserToken(User user, String token);
+}
