+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/TokenType.java
@@ -1,0 +1,5 @@
+package com.tanveer.authservice.infrastructure.persistance;
+
+public enum TokenType {
+  BEARER
+}
