+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/UserEntity.java
@@ -1,0 +1,30 @@
+package com.tanveer.authservice.infrastructure.persistance;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Entity
+@Table(name = "user_entity")
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserEntity {
+
+    @Id
+    private String id;
+
+    @Column(unique = true, nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String role;
+}
