+++ auth-service/src/main/java/com/tanveer/authservice/infrastructure/persistance/UserJpaRepository.java
@@ -1,0 +1,9 @@
+package com.tanveer.authservice.infrastructure.persistance;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserJpaRepository extends JpaRepository<UserEntity, String> {
+  Optional<UserEntity> findByEmail(String email);
+}
