--- auth-service/src/main/java/com/tanveer/authservice/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.tanveer.authservice.repository;
-
-import com.tanveer.authservice.domain.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-  Optional<UserEntity> findByEmail(String email);
-  Optional<UserEntity> findById(String id);
-
-}
