--- auth-service/src/test/java/com/tanveer/authservice/AuthServiceImplApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tanveer.authservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthServiceImplApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
