+++ auth-service/src/test/java/com/tanveer/authservice/UserServiceImplApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tanveer.authservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceImplApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
